Added files for 11 July 2018

+++ DemoLabs/DemoLab4-1.c
@@ -2,7 +2,7 @@
 
 int main (void)
 {
-    float userInput = 0; //stores user input
+    int userInput = 0; //stores user input
     printf ("Enter a character: "); //asks user for input
     userInput = getchar(); //assigns the input value (with getchar) to the variable userInput
     printf ("Your character was: "); //prefaces output
@@ -11,4 +11,3 @@ int main (void)
 
     return 0;
 }
-
